Fix data types

+++ backend/store/model.go
@@ -1,13 +1,13 @@
 package store
 
 type Profile struct {
-	Id       int64  `json:"-"`
+	Id       int32  `json:"-"`
 	Login    string `json:"login"`
 	UserName string `json:"username"`
 }
 
 type Dictionary struct {
-	Id   int64  `json:"id"`
+	Id   int32  `json:"id"`
 	Name string `json:"name"`
 }
 
@@ -16,11 +16,11 @@ type Appointment struct {
 	CreatedAt                      int64  `json:"-"`
 	UpdatedAt                      int64  `json:"-"`
 	DateReceipt                    int64  `json:"dateReceipt"`
-	DoctorId                       int64  `json:"doctorId"`
+	DoctorId                       int32  `json:"doctorId"`
 	DoctorName                     string `json:"doctorName" docx:"doctorName"`
-	PatientId                      int64  `json:"patientId"`
+	PatientId                      int32  `json:"patientId"`
 	PatientName                    string `json:"patientName" docx:"patientName"`
-	ReceiptKindId                  int64  `json:"receiptKindId"`
+	ReceiptKindId                  int32  `json:"receiptKindId"`
 	ReceiptKindName                string `json:"-"`
 	ReceiptDiagnosis               string `json:"receiptDiagnosis"`
 	Alergo                         string `json:"alergo" docx:"alergo"`
@@ -39,10 +39,10 @@ type Appointment struct {
 	ParitetSV                      string `json:"paritetSV"`
 	ParitetNB                      string `json:"paritetNB"`
 	ParitetEB                      string `json:"paritetEB"`
-	InfectionMarkersStateId        int64  `json:"infectionMarkersStateId"`
+	InfectionMarkersStateId        int32  `json:"infectionMarkersStateId"`
 	InfectionMarkersStateName      string `json:"-"`
 	InfectionMarkersDesc           string `json:"infectionMarkersDesc"`
-	TromboflebiaStateId            int64  `json:"tromboflebiaStateId"`
+	TromboflebiaStateId            int32  `json:"tromboflebiaStateId"`
 	TromboflebiaStateName          string `json:"-"`
 	TromboflebiaDesc               string `json:"tromboflebiaDesc"`
 	FirstTrimester                 string `json:"firstTrimester" docx:"firstTrimester"`
@@ -50,26 +50,26 @@ type Appointment struct {
 	ThirdTrimester                 string `json:"thirdTrimester" docx:"thirdTrimester"`
 	History                        string `json:"history" docx:"history"`
 	Oprv                           string `json:"oprv"`
-	OprvStateId                    int64  `json:"oprvStateId"`
+	OprvStateId                    int32  `json:"oprvStateId"`
 	OprvStateName                  string `json:"-"`
 	ExpByMenstruation              string `json:"expByMenstruation" docx:"expByMenstruation"`
 	ExpByFirstVisit                string `json:"expByFirstVisit" docx:"expByFirstVisit"`
 	ExpByUltraFirst                string `json:"expByUltraFirst"`
 	ExpByUltraSecond               string `json:"expByUltraSecond"`
 	ExpByUltraThird                string `json:"expByUltraThird"`
-	HealthStateId                  int64  `json:"healthStateId"`
+	HealthStateId                  int32  `json:"healthStateId"`
 	HealthStateName                string `json:"-" docx:"health"`
 	Claims                         string `json:"claims" docx:"claims"`
 	Head                           string `json:"head" docx:"headache"`
 	Vision                         string `json:"vision" docx:"vision"`
-	SkinStateId                    int64  `json:"skinStateId"`
+	SkinStateId                    int32  `json:"skinStateId"`
 	SkinStateName                  string `json:"-" docx:"skinStateName"`
 	Lymph                          string `json:"lymph" docx:"lymph"`
-	BreathStateId                  int64  `json:"breathStateId"`
+	BreathStateId                  int32  `json:"breathStateId"`
 	BreathStateName                string `json:"-" docx:"breathStateName"`
-	RaleStateId                    int64  `json:"raleStateId"`
+	RaleStateId                    int32  `json:"raleStateId"`
 	RaleStateName                  string `json:"-" docx:"raleStateName"`
-	TonesStateId                   int64  `json:"tonesStateId"`
+	TonesStateId                   int32  `json:"tonesStateId"`
 	TonesStateName                 string `json:"-" docx:"tonesStateName"`
 	Pulse                          string `json:"pulse" docx:"pulseValue"`
 	PulseType                      string `json:"pulseType" docx:"pulseType"`
@@ -81,35 +81,35 @@ type Appointment struct {
 	TongueUncoated                 bool   `json:"tongueUncoated"`
 	Throat                         string `json:"throat" docx:"throat"`
 	BellyPeriod                    string `json:"bellyPeriod" docx:"bellyPeriod"`
-	BellyStateId                   int64  `json:"bellyStateId"`
+	BellyStateId                   int32  `json:"bellyStateId"`
 	BellyStateName                 string `json:"-" docx:"bellyStateName"`
 	EpigastriumStateUse            bool   `json:"epigastriumStateUse"`
-	EpigastriumStateId             int64  `json:"epigastriumStateId"`
+	EpigastriumStateId             int32  `json:"epigastriumStateId"`
 	EpigastriumStateName           string `json:"-"`
 	ScarStateUse                   bool   `json:"scarStateUse"`
-	ScarStateId                    int64  `json:"scarStateId"`
+	ScarStateId                    int32  `json:"scarStateId"`
 	ScarStateName                  string `json:"-"`
 	Peritoneal                     string `json:"peritoneal" docx:"peritoneal"`
 	Labors                         string `json:"labors" docx:"labors"`
 	Dysuric                        bool   `json:"dysuric"`
 	Bowel                          bool   `json:"bowel"`
 	LimbSwelling                   string `json:"limbSwelling" docx:"limbSwelling"`
-	UteruseStateId                 int64  `json:"uteruseStateId"`
+	UteruseStateId                 int32  `json:"uteruseStateId"`
 	UteruseStateName               string `json:"-" docx:"uteruseStateName"`
 	FetalPositionId                int64  `json:"fetalPositionId"`
 	FetalPositionName              string `json:"-" docx:"fetalPositionName"`
-	FetalPreviaId                  int64  `json:"fetalPreviaId"`
+	FetalPreviaId                  int32  `json:"fetalPreviaId"`
 	FetalPreviaName                string `json:"-" docx:"fetalPreviaName"`
-	FetalAlignId                   int64  `json:"fetalAlignId"`
+	FetalAlignId                   int32  `json:"fetalAlignId"`
 	FetalAlignName                 string `json:"-" docx:"fetalAlignName"`
-	FetalHeartbeatId               int64  `json:"fetalHeartbeatId"`
+	FetalHeartbeatId               int32  `json:"fetalHeartbeatId"`
 	FetalHeartbeatName             string `json:"-" docx:"fetalHeartbeatName"`
-	HeartbeatRithmId               int64  `json:"heartbeatRithmId"`
+	HeartbeatRithmId               int32  `json:"heartbeatRithmId"`
 	HeartbeatRithmName             string `json:"-" docx:"heartbeatRithmName"`
 	FetalPulse                     string `json:"fetalPulse" docx:"fetalPulse"`
-	ReproductiveDischargeTypeId    int64  `json:"reproductiveDischargeTypeId"`
+	ReproductiveDischargeTypeId    int32  `json:"reproductiveDischargeTypeId"`
 	ReproductiveDischargeTypeName  string `json:"-" docx:"reproductiveDischargeTypeName"`
-	ReproductiveDischargeStateId   int64  `json:"reproductiveDischargeStateId"`
+	ReproductiveDischargeStateId   int32  `json:"reproductiveDischargeStateId"`
 	ReproductiveDischargeStateName string `json:"-" docx:"reproductiveDischargeStateName"`
 	Vdm                            string `json:"vdm" docx:"vdm"`
 	Oj                             string `json:"oj" docx:"oj"`
@@ -117,26 +117,26 @@ type Appointment struct {
 	Dcrist                         string `json:"dcrist" docx:"dcrist"`
 	Dtroch                         string `json:"dtroch" docx:"dtroch"`
 	Cext                           string `json:"cext" docx:"cext"`
-	DevelOrgansId                  int64  `json:"develOrgansId"`
+	DevelOrgansId                  int32  `json:"develOrgansId"`
 	DevelOrgansName                string `json:"-" docx:"develOrgansName"`
 	GenitalAnomalies               string `json:"genitalAnomalies" docx:"genitalAnomalies"`
-	VaginaStateId                  int64  `json:"vaginaStateId"`
+	VaginaStateId                  int32  `json:"vaginaStateId"`
 	VaginaStateName                string `json:"-" docx:"vaginaStateName"`
 	Bishop                         string `json:"bishop" docx:"bishop"`
-	FetalBladderStateId            int64  `json:"fetalBladderStateId"`
+	FetalBladderStateId            int32  `json:"fetalBladderStateId"`
 	FetalBladderStateName          string `json:"-" docx:"fetalBladderStateName"`
-	FetalBladderPreviaId           int64  `json:"fetalBladderPreviaId"`
+	FetalBladderPreviaId           int32  `json:"fetalBladderPreviaId"`
 	FetalBladderPreviaName         string `json:"-" docx:"fetalBladderPreviaName"`
-	FetalBladderAlignId            int64  `json:"fetalBladderAlignId"`
+	FetalBladderAlignId            int32  `json:"fetalBladderAlignId"`
 	FetalBladderAlignName          string `json:"-" docx:"fetalBladderAlignName"`
 	Arches                         string `json:"arches" docx:"arches"`
 	Conjugate                      string `json:"conjugate" docx:"conjugate"`
-	PelvisStateId                  int64  `json:"pelvisStateId"`
+	PelvisStateId                  int32  `json:"pelvisStateId"`
 	PelvisStateName                string `json:"-" docx:"pelvisStateName"`
 	PelvisExostosis                string `json:"pelvisExostosis" docx:"pelvisExostosis"`
-	PelvisDischargeTypeId          int64  `json:"pelvisDischargeTypeId"`
+	PelvisDischargeTypeId          int32  `json:"pelvisDischargeTypeId"`
 	PelvisDischargeTypeName        string `json:"-" docx:"pelvisDischargeTypeName"`
-	PelvisDischargeStateId         int64  `json:"pelvisDischargeStateId"`
+	PelvisDischargeStateId         int32  `json:"pelvisDischargeStateId"`
 	PelvisDischargeStateName       string `json:"-" docx:"pelvisDischargeStateName"`
 	Diagnosis                      string `json:"diagnosis" docx:"diagnosis"`
 	Conclusion                     string `json:"conclusion" docx:"conclusion"`
